Rebase of #3891 with some fixes.