@HechtiDerLachs 2024-06-30
It looks like some changes introduced here are quite expensive. `vinberg1.jlcon` did terminate, but probably took a long time. I will investigate and do some comparisons.