@HechtiDerLachs 2024-06-30
Thanks for having a look! 

I was afraid that the booktests only got cancelled, because the `vinberg_1.jlcon` had already taken up too much of the time. But indeed I have no trouble running this test locally. 

I also tried the testfile for `elliptic_surfaces.jl` and got the following timings.
Before:
```
606.089320 seconds (3.03 G allocations: 116.741 GiB, 8.48% gc time)

620.284879 seconds (3.03 G allocations: 115.920 GiB, 8.75% gc time)
```
With this branch:
```
601.933683 seconds (2.93 G allocations: 114.751 GiB, 8.59% gc time)

662.196737 seconds (2.84 G allocations: 108.817 GiB, 8.91% gc time)
```
So memory consumption decreased with no significant impact on computing time. I think this is all good then.